Consolidate image upload into shared useUploadImage hook
The two upload sites (ImageForm, InputContent) had byte-identical
mutationFns, only differing in what the caller did with the result.
Pulled the mutation into src/lib/use-upload-image.ts with the bucket,
filename convention, and cacheControl baked in — one source of truth.

Callers now attach their own side effects:
- ImageForm passes onSuccess per-call via sendImage.mutate(file, { ... }).
- InputContent awaits mutateAsync and derives the alt text from file.name
  directly (no longer shuttled through the mutation's return shape).
